Create README.md
This is an STM32L151/152 low power processor module for QFP100 packaged devices. It is a 50mm x 50mm 4 Layer PCB with two 38 pin 100mil dual row headers. It features an optional uSD card slot and ESP-07 WiFi module on-board which can be powered down individually by the processor for minimal power consumption. 

The design is done in KiCad and is released under the CERN Open Hardware License Version 1.2