STM32L151/152 Processor Module
This is an STM32L151/152 low power 32bit ARM Cortex M3 processor module. It is designed for use with the QFP100 packaged devices and brings out most IO for external use. A companion STM32CubeMX project template is also provided in the repository as well as an Atollic TrueStudio project shell. 

The processor module is on a 50mm x 50mm 4 layer PCB with two 38 pin dual row 100mil headers provided for power and I/O. There is no on-board power regulation so the board must be fed with regulated 1.8-3.3V. This module is designed primarily for battery powered applications. Internal clock speeds of up to 32MHz are supported. 

Features include a USB FS Port and optional uSD card slot and ESP-07 WiFi module. The uSD card and WiFi module may be powered up or down independently by the processor to conserve power. 

This design is released under the CERN OHL Ver 1.2